Typo: "Archaelogy" -> "Archaeology"
Verified against the printed version of the storybook.

There are still differences between the two, though. One I noticed is
that the printed version uses "mystical", not "mystick". I didn't change
that because "mystick" is a real, albeit archaic, word that seems
consistent with the "magick" and "logick" used elsewhere.